src/api/requirements.txt
attrs==23.1.0
blinker==1.6.3
cachelib==0.9.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
coverage==7.3.2
fake-useragent==1.3.0
filelock==3.13.1
Flask==3.0.0
Flask-Caching==2.1.0
Flask-Cors==4.0.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
nest-asyncio==1.5.8
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
PySocks==1.7.1
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
requests==2.31.0
requests-file==1.5.1
selenium==4.14.0
selenium-requests==2.0.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tldextract==3.5.0
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.7
Werkzeug==3.0.1
wsproto==1.2.0
